Revert "Labs hjelpetekst helligdager (#490)"
This reverts commit 5fbf87efa870f735418fed4a590e77d160060e47.
