+++ src/components/sporsmal/sporsmal-form/sporsmal-form.less
@@ -205,10 +205,6 @@
         margin: 1rem;
     }
 
-    .helligdager-tittel {
-        padding-bottom: 0.8rem;
-    }
-
     .avbrytDialog__trigger {
         margin: 0;
     }
