+++ src/components/sporsmal/typer/radio-komp.tsx
@@ -1,16 +1,12 @@
-import { Alert, BodyLong, BodyShort,Label, Radio, RadioGroup } from '@navikt/ds-react'
-import React, { useEffect, useState } from 'react'
+import { Alert, BodyLong, BodyShort } from '@navikt/ds-react'
+import React, { useEffect } from 'react'
 import { useFormContext } from 'react-hook-form'
 import useForceUpdate from 'use-force-update'
 
-import { useAppStore } from '../../../data/stores/app-store'
 import { RSSvartype } from '../../../types/rs-types/rs-svartype'
-import { rodeUkeDagerIPerioden } from '../../../utils/helligdager-utils'
 import validerArbeidsgrad from '../../../utils/sporsmal/valider-arbeidsgrad'
-import { getLedetekst,tekst } from '../../../utils/tekster'
-import { useAmplitudeInstance } from '../../amplitude/amplitude'
+import { getLedetekst, tekst } from '../../../utils/tekster'
 import AnimateOnMount from '../../animate-on-mount'
-import { Ekspanderbar } from '../../ekspanderbar/ekspanderbar'
 import FeilLokal from '../../feil/feil-lokal'
 import Vis from '../../vis'
 import { SpmProps } from '../sporsmal-form/sporsmal-form'
@@ -23,10 +19,7 @@ const RadioKomp = ({ sporsmal }: SpmProps) => {
     const watchRadio = watch(sporsmal.id)
     const feilmelding = hentFeilmelding(sporsmal)
     const forceUpdate = useForceUpdate()
-    const { valgtSoknad } = useAppStore()
-    const { validerGrad, beregnGrad } = validerArbeidsgrad(sporsmal)
-    const [ surveySvart, setSurveySvart ] = useState<boolean>(false)
-    const { logEvent } = useAmplitudeInstance()
+    const { beregnGrad } = validerArbeidsgrad(sporsmal)
 
     useEffect(() => {
         // Tvangsoppdatering for å få riktig grad i advarselboksen (NB! vi vet det er stygt)
@@ -34,7 +27,6 @@ const RadioKomp = ({ sporsmal }: SpmProps) => {
         // eslint-disable-next-line
     }, [ sporsmal ])
 
-    const lavereProsentHjelpTittel = tekst('ekspanderbarhjelp.prosenten_lavere_enn_forventet_arbeidstaker.tittel')
     return (
         <>
             <SporsmalstekstH3 sporsmal={sporsmal} />
@@ -80,7 +72,7 @@ const RadioKomp = ({ sporsmal }: SpmProps) => {
 
             <FeilLokal sporsmal={sporsmal} />
 
-            <Vis hvis={watchRadio && watchRadio.toLowerCase() === 'timer' && beregnGrad && beregnGrad() && validerGrad!() == true}
+            <Vis hvis={watchRadio && watchRadio.toLowerCase() === 'timer' && beregnGrad && beregnGrad()}
                 render={() =>
                     <Alert variant="info" style={{ marginTop: '1rem' }}>
                         <BodyShort>
@@ -90,49 +82,6 @@ const RadioKomp = ({ sporsmal }: SpmProps) => {
                             )}
                         </BodyShort>
                     </Alert>
-                }
-            />
-
-            <Vis hvis={watchRadio && watchRadio.toLowerCase() === 'timer' && validerGrad!() !== true && rodeUkeDagerIPerioden(valgtSoknad!.fom, valgtSoknad!.tom)}
-                render={() =>
-                    <>
-                        <Ekspanderbar
-                            title={lavereProsentHjelpTittel}
-                            sporsmalId={sporsmal.id}
-                            amplitudeProps={{ 'component': lavereProsentHjelpTittel, sporsmaltag: sporsmal.tag }}>
-                            <div className="avsnitt">
-                                <Label size="medium" as="h3" className="helligdager-tittel">
-                                    {tekst('ekspanderbarhjelp.helligdager.tittel')}
-                                </Label>
-                                <BodyLong spacing>{tekst('ekspanderbarhjelp.prosenten_lavere_enn_forventet_arbeidstaker.innhold1')}</BodyLong>
-                                <BodyLong spacing>{tekst('ekspanderbarhjelp.prosenten_lavere_enn_forventet_arbeidstaker.innhold2')}</BodyLong>
-                                <RadioGroup
-                                    legend={tekst('ekspanderbarhjelp.helligdager.enkelt-tittel')}
-                                    size="medium">
-                                    <Radio value={tekst('ekspanderbarhjelp.helligdager.enkelt-svar-Ja')} onClick={() => {
-                                        if (!surveySvart) {
-                                            setSurveySvart(true)
-                                            logEvent('hjelpetekst survey besvart', {
-                                                'component': tekst('ekspanderbarhjelp.helligdager.enkelt-tittel'),
-                                                'tekst': tekst('ekspanderbarhjelp.helligdager.enkelt-svar-Ja')
-                                            })
-                                        }
-                                    }
-                                    }>{tekst('ekspanderbarhjelp.helligdager.enkelt-svar-Ja')}</Radio>
-                                    <Radio value={tekst('ekspanderbarhjelp.helligdager.enkelt-svar-Nei')} onClick={() => {
-                                        if (!surveySvart) {
-                                            setSurveySvart(true)
-                                            logEvent('hjelpetekst survey besvart', {
-                                                'component': tekst('ekspanderbarhjelp.helligdager.enkelt-tittel'),
-                                                'tekst': tekst('ekspanderbarhjelp.helligdager.enkelt-svar-Nei')
-                                            })
-                                        }
-                                    }
-                                    }>{tekst('ekspanderbarhjelp.helligdager.enkelt-svar-Nei')}</Radio>
-                                </RadioGroup>
-                            </div>
-                        </Ekspanderbar>
-                    </>
                 }
             />
         </>
